requirements.txt
requests~=2.27.1
torch==1.4.0
Pillow==8.4.0
typing~=3.7.4.3
torchvision==0.11.3
opencv-python~=4.5.5.64
numpy==1.19.5
loguru~=0.6.0
uvicorn==0.17.0
fastapi~=0.78.0
starlette~=0.19.1
pydantic==1.7.4
click~=8.1.3
tqdm~=4.64.0
setuptools==59.6.0
aiofiles~=0.8.0
python-multipart~=0.0.5